Add security scan for the repository
This does a security scan for hardcoded secrets and vulnerable software
dependencies. This will hopefully help us prioritize dependency updates
in such a way that we don't keep vulnerable dependencies for too long.
Note that this should only be detecting `HIGH` and `CRITICAL`
vulnerabilities as to not overwhelm the team with warnings.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>