Add security scanning to container images
This adds a step in the GitHub Actions CI to automatically build the
krakend-ce and builder images and run a security scan on them.

Similarly to the repository scans introduced in a previous PR [1], this
uses Trivy [2] to run the scans, which allows us to add exceptions if
vulnerabilities are not relevant to these images.

[1] https://github.com/krakendio/krakend-ce/pull/650
[2] https://aquasecurity.github.io/trivy/v0.35/

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>