Sign builder images using cosign + sigstore
cosign [1] is a command line utility which will sign a specific
container image given a tag. The signature is pushed to the same
container repository in a specialized tag, and will use the public
sigstore infrastructure to host the signatures in its transparency log,
using the GitHub workflow's credentials.

To verify, do:

```
cosign verify <image reference>
```

The verification also happens from the public transparency log.

[1] https://github.com/sigstore/cosign
[2] https://www.sigstore.dev/

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>